moved AllLinesChart to separate component
